Catch failures and count 'em
I non-prod I noticed that some of the tasks with --wait are failing.

We don't want failing tasks to kill the entire job. We also don't want to bury failures. So this will keep track of the failues and mark the tasks as failing in CircleCI -- while still running as many as possible.